requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
playwright==1.48.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
jinja2==3.1.4
